# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BEOS API",
        "description": "Documentación oficial de la API BEOS",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listado paginado de productos con filtros",
                "operationId": "7c9c65dc74c69439386c99a2ced8d90b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por nombre",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency_id",
                        "in": "query",
                        "description": "Filtrar por ID de moneda",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Resultados por página",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Campo de ordenación",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "price",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Dirección",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de productos recuperado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Productos listados con éxito."
                                        },
                                        "value": {
                                            "$ref": "#/components/schemas/ProductListResponse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "728ebebe37c18429b6efc48655179ffa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "tax_cost",
                                    "currency_id",
                                    "manufacturing_cost"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Producto Premium"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 150.5
                                    },
                                    "tax_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 15
                                    },
                                    "currency_id": {
                                        "description": "ID de la moneda existente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "manufacturing_cost": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 80
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción detallada del producto",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto guardado correctamente."
                                        },
                                        "value": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Producto Premium"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descripción..."
                                                },
                                                "price": {
                                                    "type": "string",
                                                    "example": "150.50"
                                                },
                                                "currency": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Dólar"
                                                        },
                                                        "symbol": {
                                                            "type": "string",
                                                            "example": "USD"
                                                        },
                                                        "exchange_rate": {
                                                            "type": "string",
                                                            "example": "1.00"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "manufacturing_cost": {
                                                    "type": "string",
                                                    "example": "80.00"
                                                },
                                                "tax_cost": {
                                                    "type": "string",
                                                    "example": "15.00"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-03-20"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-03-20"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación (Ej. Nombre duplicado o campos faltantes)"
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hubo un error al procesar la solicitud."
                                        },
                                        "value": {
                                            "type": "string",
                                            "example": "Mensaje de la excepción..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Producto A"
                    },
                    "price": {
                        "type": "string",
                        "example": "100.00"
                    }
                },
                "type": "object"
            },
...
            "ProductListResponse": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductResource"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Ingresa solo el token (ej: 4|lGMHH...) y Swagger añadirá \"Bearer \" por ti."
            }
        }
    },
...
}